requirements.txt
asgiref==3.7.2
Django==4.2.5
django-countries==7.5.1
django-formtools==2.4.1
sqlparse==0.4.4
typing_extensions==4.8.0
